pyproject.toml
[tool.poetry]
name = "bookstore"
version = "0.1.0"
description = "Bookstore API"
authors = ["Flavio Carvalho <professorflaviosoares@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
pytest = "^8.3.2"
factory-boy = "^3.3.0"
django = "^5.0.7"
django-extensions = "^3.2.3"
django-debug-toolbar = "^4.4.6"
djangorestframework = "^3.14.0"  # Adicione esta linha
psycopg2-binary = "^2.9.9"
flake8 = "^7.1.1"

[tool.poetry.group.dev.dependencies]
pytest-django = "^4.9.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
